requirements.txt.txt
python>=3.11
PyPDF2==3.0.0
transformers==4.30.0
numpy==1.23.5
scikit-learn==1.1.2
sentence-transformers==2.2.2
rapidfuzz==2.13.7
keras==2.11.0
pydub==0.25.1
requests==2.28.1
ipywidgets==8.0.4
